+++ src/mirrors/MirrorLoader.ts
@@ -11,49 +11,30 @@ import { getMyMangaReaderCMSMirrors } from "./sites/MyMangaReaderCMS/MyMangaRead
 import { getReadMangaMirrors } from "./sites/ReadManga/ReadManga"
 
 import { MirrorHelper } from "./MirrorHelper"
-import { Manga4Life } from "./sites/Manga4Life"
 import { MangaFox } from "./sites/MangaFox"
 import { MangaHub } from "./sites/MangaHub"
 import { MangaHere } from "./sites/MangaHere"
 import { WebToon } from "./sites/WebToon"
 import { ZeroScans } from "./sites/ZeroScans"
 import { MangadexV5 } from "./sites/Mangadex-V5"
-import { ZaHardTop } from "./sites/ZaHardTop"
-import { UnionLeitor } from "./sites/UnionLeitor"
 import { Tachidesk } from "./sites/Tachidesk"
-import { SeriManga } from "./sites/SeriManga"
-import { ScyllaScans } from "./sites/ScyllaScans"
 import { ScanTradUnion } from "./sites/ScanTradUnion"
 import { SadScans } from "./sites/SadScans"
-import { ReadM } from "./sites/ReadM"
 import { ReadComicOnline } from "./sites/ReadComicOnline"
 import { Niceoppai } from "./sites/Niceoppai"
 import { NHentai } from "./sites/NHentai"
 import { Manhwa18com } from "./sites/Manhwa18.com"
-import { ManhwaFreak } from "./sites/ManhwaFreak"
-import { MangaSee } from "./sites/MangaSee"
 import { MangaPill } from "./sites/MangaPill"
-import { MangaDemon } from "./sites/MangaDemon"
 import { MangaKawaii } from "./sites/MangaKawaii"
 import { MangaKatana } from "./sites/MangaKatana"
-import { MangaHasu } from "./sites/MangaHasu"
-import { MangaFreak } from "./sites/MangaFreak"
-import { MangaAlArab } from "./sites/MangaAlArab"
-import { LugnicaScan } from "./sites/LugnicaScan"
 import { Komga } from "./sites/Komga"
 import { FMTeam } from "./sites/FMTeam"
-import { EpsilonScan } from "./sites/EpsilonScan"
-import { BananaScan } from "./sites/BananaScan"
 import { BatotoFake } from "./sites/Batoto-fake"
 import { DynastyScans } from "./sites/DynastyScans"
-import { DisasterScans } from "./sites/DisasterScans"
 import { Dm5 } from "./sites/Dm5"
-import { ReaperScans } from "./sites/ReaperScans"
 import { LikeManga } from "./sites/LikeManga"
 import { MangaPark } from "./sites/MangaPark"
 import { MangaReaderTo } from "./sites/MangaReaderTo"
-import { FanComics } from "./sites/FanComics"
-import { MangaBTT } from "./sites/MangaBTT"
 import { MangaBuddy } from "./sites/MangaBuddy"
 import { TestSite } from "./sites/_localTestServer"
 
@@ -111,45 +92,26 @@ export const getMirrorLoader = (mirrorHelper: MirrorHelper) => {
             new WebToon(mirrorHelper),
             new MangaFox(mirrorHelper),
             new MangaHere(mirrorHelper),
-            new Manga4Life(mirrorHelper),
             new MangaHub(mirrorHelper),
             new ZeroScans(mirrorHelper),
-            new ZaHardTop(mirrorHelper),
-            new UnionLeitor(mirrorHelper),
             new Tachidesk(mirrorHelper),
-            new SeriManga(mirrorHelper),
-            new ScyllaScans(mirrorHelper),
             new ScanTradUnion(mirrorHelper),
             new SadScans(mirrorHelper),
-            new ReadM(mirrorHelper),
             new ReadComicOnline(mirrorHelper),
             new Niceoppai(mirrorHelper),
             new NHentai(mirrorHelper),
             new Manhwa18com(mirrorHelper),
-            new ManhwaFreak(mirrorHelper),
-            new MangaSee(mirrorHelper),
             new MangaPill(mirrorHelper),
-            new MangaDemon(mirrorHelper),
             new MangaKawaii(mirrorHelper),
             new MangaKatana(mirrorHelper),
-            new MangaHasu(mirrorHelper),
-            new MangaFreak(mirrorHelper),
-            new MangaAlArab(mirrorHelper),
-            new LugnicaScan(mirrorHelper),
             new Komga(mirrorHelper),
             new FMTeam(mirrorHelper),
-            new EpsilonScan(mirrorHelper),
-            new BananaScan(mirrorHelper),
             new BatotoFake(mirrorHelper),
             new DynastyScans(mirrorHelper),
-            new DisasterScans(mirrorHelper),
             new Dm5(mirrorHelper),
-            new ReaperScans(mirrorHelper),
             new LikeManga(mirrorHelper),
             new MangaPark(mirrorHelper),
             new MangaReaderTo(mirrorHelper),
-            new FanComics(mirrorHelper),
-            new MangaBTT(mirrorHelper),
             new MangaBuddy(mirrorHelper),
             new TestSite(mirrorHelper),