+++ src/mirrors/sites/MyMangaReaderCMS/MyMangaReaderCMS.ts
@@ -118,36 +118,6 @@ class MyMangaReaderCMS extends BaseMirror implements MirrorImplementation {
 }
 
 export const getMyMangaReaderCMSMirrors = (mirrorHelper: MirrorHelper): MirrorImplementation[] => {
-    return [
-        new MyMangaReaderCMS(
-            mirrorHelper,
-            {
-                mirrorName: "Komikid",
-                mirrorIcon: require("../../icons/id-optimized.png"),
-                languages: "id",
-                domains: ["komikid.com"],
-                home: "http://www.komikid.com",
-                chapter_url: /^\/manga\/.*\/.+$/g
-            },
-            {
-                base_url: "http://www.komikid.com",
-                chapters_element: "ul.chapters > li > h5 > a"
-            }
-        ),
-        new MyMangaReaderCMS(
-            mirrorHelper,
-            {
-                mirrorName: "ScanFR",
-                mirrorIcon: require("../../icons/scanfr-optimized.png"),
-                languages: "fr",
-                domains: ["scan-fr.cc", "scan-fr.org"],
-                home: "https://www.scan-fr.cc",
-                chapter_url: /^\/manga\/.*\/.+$/g
-            },
-            {
-                base_url: "https://www.scan-fr.org",
-                chapters_element: "ul.chapterszozo a[href*='/manga/']"
-            }
-        )
-    ]
+    // All MyMangaReaderCMS implementations have been removed as the sites are dead
+    return []
 }